requirements.txt
alembic==0.8.9
asn1crypto==0.24.0
autopep8==1.4.3
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
coverage==4.5.3
cryptography==2.6.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.2
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.1.4
Werkzeug==0.14.1
